refactor(frontend): WorkspaceView 拆分 — 提取 SidebarNav 组件 + 接入 composable
- 新增 SidebarNav.vue（PC 侧边栏 + 移动端底部导航）
- 新增 useWorkspaceNav（视图路由/导航配置）
- 新增 useChatSession（AI 辅导对话）
- 接入已有 useFileUpload、useSplitPipeline
- WorkspaceView 从 1372 行缩减至 ~830 行
